+++ .travis.yml
@@ -3,9 +3,6 @@ sudo: false
 
 matrix:
   include:
-    - go: 1.4
-    - go: 1.5.x
-    - go: 1.6.x
     - go: 1.7.x
     - go: 1.8.x
     - go: 1.9.x
